Make Quick Start more approachable
At DragonSprint we sent people over the Flask front page. Then we sent them over
the TryPyramid front page. Both “above the fold“. The difference was staggering:
all managed to start the Flask example and got a browser to render “Hello
World“. And none managed to do it with Pyramid without help or frustration.

This is our attempt to make TryPyramid QuickStart example more approachable.

Change 1: the Quick Start code configures the view imperatively and the Pyramid
Feature sections provide declarative examples. Lets promote a single way of view configuration, at least for newcomers.

Change 2: Every single person who tried the Quick Start code today got stuck
with a 404 Not Found error when running the code and opening it in the browser.
Because one needs to go to ``/hello/foo`` to get a response. This is not obvious
at all and people get stuck. Also, while it is nice to show off some Pyramid
features, let’s keep this example as simple as possible and return a hardcoded
string, on root. Several less things to get stuck on.